moved get_cell_df to mibi_data
