+++ deepcell/tmp_funcs.py
@@ -120,70 +120,6 @@ def load_celltypes(path, point_list):
 Data handeling
 """
 
-def get_cell_df(mibi_data, mibi_labels, mibi_celltypes, markers, marker_idx_dict):
-    """Create a data frame that has all cell information in one place
-
-    Args:
-        mibi_data (np.array):
-        mibi_labels (np.array):
-        mibi_celltypes (np.array):
-        markers (list):
-        marker_idx_dict (dictionary):
-
-    Returns:
-        pd.DataFrame (float): cell by feature matrix with batch info and cell type
-
-    """
-    num_batches = mibi_data.shape[0]
-    num_feats = mibi_data.shape[-1]
-
-    cell_df = pd.DataFrame(0.0, index=range(0), columns=range(num_feats + 2))
-    col_names = np.copy(markers).tolist()
-    col_names.insert(0, 'label')
-    col_names.insert(0, 'batch')
-    cell_df.columns = col_names
-    cell_df['cell_type'] = np.nan
-
-    for batch in range(num_batches):
-        mibi_image = mibi_data[batch]
-        label_image = np.squeeze(mibi_labels[batch])
-        type_image = np.squeeze(mibi_celltypes[batch])
-
-        props = regionprops(label_image)
-
-        num_cells = len(props)
-        num_feats = mibi_data.shape[-1]
-
-        batch_df = pd.DataFrame(0.0, index=range(num_cells), columns=range(num_feats + 2))
-        batch_df.rename(columns={0:'batch', 1:'label'}, inplace = True)
-
-        batch_df['batch'] = batch
-
-        for feature in range(num_feats):
-            marker = marker_idx_dict[feature]
-
-            props_table = regionprops_table(label_image,
-                                            intensity_image=mibi_image[..., feature],
-                                            properties=['mean_intensity', 'label'])
-
-            batch_df['label'] = props_table['label']
-            batch_df[feature + 2] = props_table['mean_intensity']
-            batch_df.rename(columns={feature+2: marker}, inplace=True)
-
-        batch_df['cell_type'] = np.nan
-
-        for prop in props:
-            idx = batch_df[batch_df['label'] == prop.label].index[0]
-
-            coords = prop.coords
-            cts = type_image[coords[:, 0], coords[:, 1]]
-            cell_type = mode(cts, axis=None).mode[0]
-
-            batch_df.loc[idx, 'cell_type'] = cell_type
-
-        cell_df = pd.concat([cell_df, batch_df])
-
-    return cell_df
 
 
 def qt_transform(X):
